Django/DMS/requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
Django==4.0.5
djangorestframework==3.13.1
idna==3.3
mariadb==1.0.11
paho-mqtt==1.6.1
PyMySQL==1.0.2
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
scikit-learn==1.1.1
pandas==1.4.2
